consolidate current working directory; considate helpers scripts
