+++ scripts/add_electricity.py
@@ -803,9 +803,8 @@ if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake("add_electricity")
-        os.chdir(snakemake.config["ROOT_PATH"])
+
     configure_logging(snakemake)
 
     n = pypsa.Network(snakemake.input.base_network)
