+++ scripts/add_existing_baseyear.py
@@ -579,9 +579,8 @@ def add_heating_capacities_installed_before_baseyear(
 
 if __name__ == "__main__":
     if "snakemake" not in globals():
-        from helpers import mock_snakemake
+        from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake(
             "add_existing_baseyear",
             simpl="",
