+++ scripts/add_extra_components.py
@@ -270,8 +270,8 @@ if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake("add_extra_components", simpl="", clusters=10)
+
     configure_logging(snakemake)
 
     n = pypsa.Network(snakemake.input.network)
