+++ scripts/augmented_line_connections.py
@@ -54,10 +54,10 @@ if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake(
             "augmented_line_connections", network="elec", simpl="", clusters="54"
         )
+
     configure_logging(snakemake)
 
     n = pypsa.Network(snakemake.input.network)
