+++ scripts/build_bus_regions.py
@@ -150,8 +150,8 @@ if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake("build_bus_regions")
+
     configure_logging(snakemake)
 
     countries = snakemake.params.countries
