+++ scripts/build_cop_profiles.py
@@ -26,7 +26,7 @@ def coefficient_of_performance(delta_T, source="air"):
 
 if __name__ == "__main__":
     if "snakemake" not in globals():
-        from helpers import mock_snakemake
+        from _helpers import mock_snakemake
 
         snakemake = mock_snakemake(
             "build_cop_profiles",
