+++ scripts/build_cutout.py
@@ -107,8 +107,8 @@ if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake("build_cutout", cutout="africa-2013-era5")
+
     configure_logging(snakemake)
 
     cutout_params = snakemake.params.cutouts[snakemake.wildcards.cutout]
