+++ scripts/build_demand_profiles.py
@@ -294,9 +294,8 @@ if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake("build_demand_profiles")
-        os.chdir(snakemake.config["ROOT_PATH"])
+
     configure_logging(snakemake)
 
     n = pypsa.Network(snakemake.input.base_network)
