+++ scripts/build_powerplants.py
@@ -298,7 +298,6 @@ if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake("build_powerplants")
 
     configure_logging(snakemake)
