+++ scripts/build_renewable_profiles.py
@@ -488,9 +488,7 @@ if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake("build_renewable_profiles", technology="solar")
-        os.chdir(snakemake.config["ROOT_PATH"])
     configure_logging(snakemake)
 
     pgb.streams.wrap_stderr()
