+++ scripts/build_solar_thermal_profiles.py
@@ -16,15 +16,13 @@ import xarray as xr
 
 if __name__ == "__main__":
     if "snakemake" not in globals():
-        from helpers import mock_snakemake, sets_path_to_root
+        from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake(
             "build_solar_thermal_profiles",
             simpl="",
             clusters=15,
         )
-        os.chdir(snakemake.config["ROOT_PATH"])
 
     config = snakemake.params.solar_thermal_config
