+++ scripts/build_temperature_profiles.py
@@ -15,15 +15,13 @@ import xarray as xr
 
 if __name__ == "__main__":
     if "snakemake" not in globals():
-        from helpers import mock_snakemake, sets_path_to_root
+        from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake(
             "build_temperature_profiles",
             simpl="",
             clusters=900,
         )
-        os.chdir(snakemake.config["ROOT_PATH"])
 
     time = pd.date_range(freq="h", **snakemake.params.snapshots)
     cutout_path = (
