+++ scripts/build_test_configs.py
@@ -88,7 +88,6 @@ if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake("build_test_configs")
 
     # Input paths
