+++ scripts/clean_osm_data.py
@@ -1064,7 +1064,6 @@ if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake("clean_osm_data")
     configure_logging(snakemake)
