+++ scripts/copy_config.py
@@ -14,7 +14,7 @@ files_to_copy = {
 
 if __name__ == "__main__":
     if "snakemake" not in globals():
-        from helpers import mock_snakemake
+        from _helpers import mock_snakemake
 
         snakemake = mock_snakemake("copy_config")
