+++ scripts/download_osm_data.py
@@ -94,9 +94,7 @@ if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake("download_osm_data")
-        os.chdir(snakemake.config["ROOT_PATH"])
     configure_logging(snakemake)
 
     run = snakemake.config.get("run", {})
