+++ scripts/monte_carlo.py
@@ -350,7 +350,6 @@ if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake(
             "monte_carlo",
             simpl="",
