+++ scripts/prepare_db.py
@@ -31,7 +31,7 @@ import pypsa
 
 if __name__ == "__main__":
     if "snakemake" not in globals():
-        from helpers import mock_snakemake
+        from _helpers import mock_snakemake
 
         snakemake = mock_snakemake(
             "prepare_db",
