+++ scripts/prepare_transport_data_input.py
@@ -12,7 +12,7 @@ import helpers
 import numpy as np
 import pandas as pd
 
-# from helpers import configure_logging
+# from _helpers import configure_logging
 
 
 # logger = logging.getLogger(__name__)
@@ -109,11 +109,10 @@ def download_CO2_emissions():
 
 if __name__ == "__main__":
     if "snakemake" not in globals():
-        from helpers import mock_snakemake, sets_path_to_root
+        from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake("prepare_transport_data_input")
-        os.chdir(snakemake.config["ROOT_PATH"])
+
     # configure_logging(snakemake)
 
     # run = snakemake.config.get("run", {})
