+++ scripts/simplify_network.py
@@ -964,8 +964,8 @@ if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
-        os.chdir(os.path.dirname(os.path.abspath(__file__)))
         snakemake = mock_snakemake("simplify_network", simpl="")
+
     configure_logging(snakemake)
 
     n = pypsa.Network(snakemake.input.network)
